Move `crate_dir` into method on `ExCrate`.
